mood_lens/requirements.txt
accelerate==1.12.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
attrs==25.4.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
coloredlogs==15.0.1
contourpy==1.3.3
cycler==0.12.1
faiss-cpu==1.13.0
fastapi==0.123.4
filelock==3.20.0
flatbuffers==25.9.23
fonttools==4.61.0
frozenlist==1.8.0
fsspec==2025.10.0
fvcore==0.1.5.post20221221
genai==2.1.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
humanfriendly==10.0
idna==3.11
iopath==0.1.10
ipython==8.37.0
Jinja2==3.1.6
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.7
mpmath==1.3.0
multidict==6.7.0
networkx==3.6
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
ollama==0.6.1
onnxruntime==1.23.2
onnxruntime-gpu==1.23.2
openai==0.27.10
opencv-python-headless==4.12.0.88
pandas==2.3.3
peft==0.18.0
pillow==12.0.0
portalocker==3.2.0
propcache==0.4.1
protobuf==6.33.1
pydantic==2.12.5
pydantic_core==2.41.5
pyparsing==3.2.5
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
safetensors==0.7.0
seaborn==0.13.2
starlette==0.50.0
sympy==1.14.0
tabulate==0.9.0
termcolor==3.2.0
tiktoken==0.3.3
tokenizers==0.22.1
torch==2.9.1
torchvision==0.24.1
tqdm==4.67.1
transformers==4.57.3
triton==3.5.1
typing-inspection==0.4.2
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.38.0
yacs==0.1.8
yarl==1.22.0
